requirements.txt
face_recognition==1.3.0
Click==6.0
dlib==19.3.0
numpy
Pillow==9.2.8
scipy==0.17.0
opencv-python==4.6.0.66

pip install numpy
pip install opencv-python

pip install pillow
pip install dlib

pip install click
pip install scipy